lyra_transform.ts へのコード分離は不要（worker.js のサイズに影響なし）だったのファイル削除
